Move core types into single file
